pyexpat is only used when updating, move import inside class.
Update is not intended to be used on Ben NanoNote, so I'm moving the
import of pyexpat to the Updater class.  pyexpat is not included in
the default build, and this patch saves me from having to install
it.

I will in the future include pyexpat and enable full support for
generating new dump files, but this patch will still function in
that case.