move makfa.dump from /var/lib/makfa to /usr/share/makfa
I originally choose /var/lib/... because makfa.dump can be updated by
the makfa program.  /usr/share/... is for read-only data, which this
most certainly is not.

The best solution for this problem would be to have some kind of
/ovelay/var/... directory that is copied at boot to /var/...  This
would allow files that need to be persistent in /var/... to be so,
but would permit the flexibility currently seen for I/O heavy uses
of /var/...

Ater a discussion with xiangfu, we decided that modifying the
base-package to support an overlay on /var is too much work to do
right now, and that we would use /usr/share/... for the time being.