makfa: only import ElementTree if we're going to use it.
ElementTree is not available by default.  (it is in pyexpat, which is
a non-default module.)

since ElementTree is only used by the Updater code, and this code is
not designed to run on the NanoNote (it is only run as part of the
build process), move the import statement inside the Updater in the
hopes that the user on the NanoNote never has to deal with it.

In the future I will make pyexpat a default module for the Lojban
config, but this change will still see a performance gain in the
most typical use case for makfa.