netsurf: update to 2.9
Well, this is the huge update of netsurf libraries and netsurf itself.
Previously, SDL interface wasn't usable (not without keymouse), because
all netsurf keybindings only work in gtk (or something like that). At
some point of time (before this commit), netsurf wasn't buildable at
all. Now it at least builds and shows up the interface on Ben. However,
there are two problems: it crashes when i press arrow keys and
it "fails to parse" an HTML page.
So i'm commiting this because from the quality of point view, netsurf package
hasn't become worse (it's compilable and somewhat startable) - so
somebody who is interested may contact netsurf developers to report bugs
or ask for a help.